third_party/blink/web_tests/external/wpt/badging/setAppBadge_cross_origin.sub.https-expected.txt

This is a testharness.js-based test.
[PASS] Test that navigator.setAppBadge is available
[FAIL] Test that calling setAppBadge in a cross-origin iframe throws a SecurityError
  assert_equals: setAppBadge should have rejected with an error expected "error" but got "success"
[PASS] Test that calling setAppBadge in a same-origin iframe succeeds
Harness: the test ran to completion.
